+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 0.8.0
+## 0.8.3
 
+* Updates inapp module dependency to 0.8.3
+* Add support for versioned providers
+* Update [BREAKING] session init handler
 * Updates the UI with a verification review banner in the verification flow
-* Updates inapp module dependency to 0.8.0
 
 ## 0.7.0